cloud-logging/src/requirements.txt
Flask==2.0.3
google-cloud-logging==3.0.0